requirement.txt
Django==1.11
django-filter==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
Markdown==2.6.11
mongoengine==0.15.0
pymongo==3.6.0
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
